server/requirements.txt
astroid==2.3.3
blinker==1.4
certifi==2020.4.5.1
click==7.1.1
dnspython==1.16.0
dominate==2.5.1
enum-compat==0.0.3
enum34==1.1.10
eventlet==0.25.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SocketIO==4.2.1
greenlet==0.4.15
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
logilab-common==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
parso==0.6.2
pylint==2.4.4
python-dotenv==0.12.0
python-engineio==3.12.1
python-socketio==4.5.1
redis==3.4.1
sentry-sdk==0.14.3
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
visitor==0.1.3
WebOb==1.8.6
Werkzeug==1.0.1
wrapt==1.12.1
